+++ .gitmodules
@@ -1,6 +1,7 @@
 [submodule "deno_third_party"]
 	path = third_party
 	url = https://github.com/denoland/deno_third_party.git
+	shallow = true
 [submodule "std/wasi/testdata"]
 	path = std/wasi/testdata
 	url = https://github.com/khronosproject/wasi-test-suite.git
