+++ cli/dts/lib.deno.unstable.d.ts
@@ -318,6 +318,20 @@ declare namespace Deno {
     /** Import emit helpers (e.g. `__extends`, `__rest`, etc..) from
      * [tslib](https://www.npmjs.com/package/tslib). */
     importHelpers?: boolean;
+    /** This flag controls how `import` works, there are 3 different options:
+     * 
+     * - `remove`: The default behavior of dropping import statements which only
+     *   reference types.
+     * - `preserve`: Preserves all `import` statements whose values or types are
+     *   never used. This can cause imports/side-effects to be preserved.
+     * - `error`: This preserves all imports (the same as the preserve option),
+     *   but will error when a value import is only used as a type. This might
+     *   be useful if you want to ensure no values are being accidentally
+     *   imported, but still make side-effect imports explicit.
+     * 
+     * This flag works because you can use `import type` to explicitly create an
+     * `import` statement which should never be emitted into JavaScript. */
+    importsNotUsedAsValues?: "remove" | "preserve" | "error";
     /** Emit a single file with source maps instead of having a separate file.
      * Defaults to `false`. */
     inlineSourceMap?: boolean;
