+++ cli/module_graph.rs
@@ -952,6 +952,7 @@ impl Graph {
           let extension = match emitted_file.media_type {
             MediaType::JavaScript => ".js",
             MediaType::SourceMap => ".js.map",
+            MediaType::Dts => ".d.ts",
             _ => unreachable!(),
           };
           let key = format!("{}{}", specifier, extension);
@@ -1959,6 +1960,24 @@ pub mod tests {
   }
 
   #[tokio::test]
+  async fn test_graph_check_ignores_dynamic_import_errors() {
+    let specifier =
+      ModuleSpecifier::resolve_url_or_path("file:///tests/dynamicimport.ts")
+        .expect("could not resolve module");
+    let (graph, _) = setup(specifier).await;
+    let result_info = graph
+      .check(CheckOptions {
+        debug: false,
+        emit: false,
+        lib: TypeLib::DenoWindow,
+        maybe_config_path: None,
+        reload: false,
+      })
+      .expect("should have checked");
+    assert!(result_info.diagnostics.is_empty());
+  }
+
+  #[tokio::test]
   async fn fix_graph_check_emit_diagnostics() {
     let specifier =
       ModuleSpecifier::resolve_url_or_path("file:///tests/diag.ts")
@@ -2003,6 +2022,27 @@ pub mod tests {
     assert!(result_info.diagnostics.is_empty());
     let h = handler.borrow();
     assert_eq!(h.cache_calls.len(), 0);
+    assert_eq!(h.tsbuildinfo_calls.len(), 1);
+  }
+
+  #[tokio::test]
+  async fn fix_graph_check_mjs_root() {
+    let specifier = ModuleSpecifier::resolve_url_or_path("file:///tests/a.mjs")
+      .expect("could not resolve module");
+    let (graph, handler) = setup(specifier).await;
+    let result_info = graph
+      .check(CheckOptions {
+        debug: false,
+        emit: true,
+        lib: TypeLib::DenoWindow,
+        maybe_config_path: None,
+        reload: false,
+      })
+      .expect("should have checked");
+    assert!(result_info.maybe_ignored_options.is_none());
+    assert!(result_info.diagnostics.is_empty());
+    let h = handler.borrow();
+    assert_eq!(h.cache_calls.len(), 1);
     assert_eq!(h.tsbuildinfo_calls.len(), 1);
   }
 
@@ -2141,6 +2181,50 @@ pub mod tests {
     let actual = actual.unwrap();
     assert!(actual.contains("const b = \"b\";"));
     assert!(actual.contains("console.log(b);"));
+  }
+
+  #[tokio::test]
+  async fn fix_graph_emit_declaration() {
+    let specifier =
+      ModuleSpecifier::resolve_url_or_path("file:///a.ts").unwrap();
+    let graph = setup_memory(
+      specifier,
+      map!(
+        "/a.ts" => r#"
+        import * as b from "./b.ts";
+
+        console.log(b);
+      "#,
+        "/b.ts" => r#"
+        export const b = "b";
+      "#
+      ),
+    )
+    .await;
+    let mut user_config = HashMap::<String, Value>::new();
+    user_config.insert("declaration".to_string(), json!(true));
+    let (emitted_files, result_info) = graph
+      .emit(EmitOptions {
+        bundle_type: BundleType::None,
+        debug: false,
+        maybe_user_config: Some(user_config),
+      })
+      .expect("should have emitted");
+    assert!(result_info.diagnostics.is_empty());
+    assert!(result_info.maybe_ignored_options.is_none());
+    assert_eq!(emitted_files.len(), 6);
+    let out_a = emitted_files.get("file:///a.ts.js");
+    assert!(out_a.is_some());
+    let out_a = out_a.unwrap();
+    assert!(out_a.starts_with("import * as b from"));
+    assert!(emitted_files.contains_key("file:///a.ts.js.map"));
+    assert!(emitted_files.contains_key("file:///a.ts.d.ts"));
+    let out_b = emitted_files.get("file:///b.ts.js");
+    assert!(out_b.is_some());
+    let out_b = out_b.unwrap();
+    assert!(out_b.starts_with("export const b = \"b\";"));
+    assert!(emitted_files.contains_key("file:///b.ts.js.map"));
+    assert!(emitted_files.contains_key("file:///b.ts.d.ts"));
   }
 
   #[tokio::test]
