+++ cli/tests/module_graph/file_tests-a.mjs
@@ -1,0 +1,3 @@
+import * as b from "./b.ts";
+
+console.log(b);
