+++ cli/tests/module_graph/file_tests-b.ts
@@ -1,0 +1,1 @@
+export const b = "b";
