+++ cli/tests/module_graph/file_tests-dynamicimport.ts
@@ -1,0 +1,5 @@
+try {
+  await import("bare_module_specifier");
+} catch (err) {
+  console.log(err);
+}
