+++ std/node/_crypto.ts
@@ -1,0 +1,3 @@
+// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+export { default as randomBytes } from "./_crypto/randomBytes.ts";
+export { pbkdf2, pbkdf2Sync } from "./_crypto/pbkdf2.ts";
