+++ std/node/_crypto/types.ts
@@ -1,3 +1,3 @@
-import Buffer from "../buffer.ts";
+import { Buffer } from "../buffer.ts";
 
 export type HASH_DATA = string | ArrayBufferView | Buffer;
