+++ std/node/path.ts
@@ -1,2 +1,4 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 export * from "../path/mod.ts";
+import * as m from "../path/mod.ts";
+export default m;
